Split Route Planner Menu into smaller components
